Add node django-test.
This was successfully bootstrapped with this command:

    knife bootstrap IP_ADDRESS \
        --distro=debian-vanilla-apt \
        --environment=test \
        --node-name=django-test \
        --ssh-user=root \
        --ssh-password=PASSWORD
        --run-list="role[default]"

The order of the options seems to matter since swapping the
positions of --environment and --node-name caused chef-client to
try and authenticate as the old hostname rather than the chosen
node name.

The django-test.js file was written manually to match the
arguments provided to knife bootstrap, but it should be possible
to instead download settings created on the server and save them.