Put hostname recipe after autoetchosts.
This seems to work seems hostname is friendlier about leaving things
in the hosts file that were placed by other things, unlike
autoetchosts which just overwrites the whole thing.

Interestingly, it seemed to take a couple iterations of chef-client
on either node before converging on a hosts file that had all nodes
and all FQDNs.